requirements.txt
asyncio
discord.py
praw
aiosqlite
dblpy